app/requirements.txt
django
django-ninja
django-environ
psycopg2
django-cors-headers